+++ packages/play/src/pages/DatePickerDemo.vue
@@ -1,10 +1,20 @@
-<script lang="ts" setup></script>
+<script lang="ts" setup>
+import { reactive } from 'vue'
+
+const data = reactive({
+  a: ''
+})
+</script>
 <template>
-  <zl-form>
+  <zl-form :data="data">
     <zl-row :columns="2">
       <zl-col>
         <zl-form-item prop="s" label="日期选择器">
-          <zl-date-picker format="YYYYMMDD" weekStart="monday"></zl-date-picker>
+          <zl-date-picker
+            :clearable="true"
+            format="YYYYMMDD"
+            weekStart="monday"
+            v-model="data.a"></zl-date-picker>
         </zl-form-item>
       </zl-col>
     </zl-row>