+++ changelog.md
@@ -69,3 +69,4 @@
 ### feature/v0.0.18 需要处理的问题
 
 1. 新增radio组件
+2. datePicker增加reset方法，可以一键清空按钮